# Use Linesweep for Linestring IsSimple