The main part of this change is to use a linesweep algorithm for LineString IsSimple (rather than exhaustively checking for intersections between each pair of line segments).

Benchmark numbers are a bit noisy, but in particular show significantly better behaviour (time and memory usage) for `BenchmarkPolygonSingleRingValidation`. This benchmark is just testing if a polygon with a single ring (made of an increasing number of line segments) is valid or not. Internally, this uses `IsSimple` on the polygon rings.

I haven't done any performance analysis using pprof yet, so there are likely some wins still that can be made WRT micro-optimisations.

```
benchmark                                               old ns/op       new ns/op       delta
BenchmarkIntersectsLineStringWithLineString/n=10        2406            2297            -4.53%
BenchmarkIntersectsLineStringWithLineString/n=100       22914           22902           -0.05%
BenchmarkIntersectsLineStringWithLineString/n=1000      331670          293550          -11.49%
BenchmarkIntersectsLineStringWithLineString/n=10000     3339903         3382322         +1.27%
BenchmarkPolygonSingleRingValidation/n=10               8505            7644            -10.12%
BenchmarkPolygonSingleRingValidation/n=100              790249          108959          -86.21%
BenchmarkPolygonSingleRingValidation/n=1000             81407496        979844          -98.80%
BenchmarkPolygonSingleRingValidation/n=10000            8306103613      10869523        -99.87%
BenchmarkPolygonMultipleRingsValidation/n=4             57655           63350           +9.88%
BenchmarkPolygonMultipleRingsValidation/n=36            2343242         2448954         +4.51%
BenchmarkPolygonMultipleRingsValidation/n=400           266714522       271292225       +1.72%
BenchmarkPolygonMultipleRingsValidation/n=4096          32014830210     31947148328     -0.21%

benchmark                                               old allocs     new allocs     delta
BenchmarkIntersectsLineStringWithLineString/n=10        17             17             +0.00%
BenchmarkIntersectsLineStringWithLineString/n=100       17             17             +0.00%
BenchmarkIntersectsLineStringWithLineString/n=1000      17             17             +0.00%
BenchmarkIntersectsLineStringWithLineString/n=10000     17             17             +0.00%
BenchmarkPolygonSingleRingValidation/n=10               158            99             -37.34%
BenchmarkPolygonSingleRingValidation/n=100              14966          900            -93.99%
BenchmarkPolygonSingleRingValidation/n=1000             1499519        8907           -99.41%
BenchmarkPolygonSingleRingValidation/n=10000            149995029      89704          -99.94%
BenchmarkPolygonMultipleRingsValidation/n=4             909            934            +2.75%
BenchmarkPolygonMultipleRingsValidation/n=36            37277          37462          +0.50%
BenchmarkPolygonMultipleRingsValidation/n=400           4126635        4128640        +0.05%
BenchmarkPolygonMultipleRingsValidation/n=4096          428296227      428316712      +0.00%

benchmark                                               old bytes       new bytes       delta
BenchmarkIntersectsLineStringWithLineString/n=10        1152            1152            +0.00%
BenchmarkIntersectsLineStringWithLineString/n=100       6976            6976            +0.00%
BenchmarkIntersectsLineStringWithLineString/n=1000      66112           66112           +0.00%
BenchmarkIntersectsLineStringWithLineString/n=10000     655936          655936          +0.00%
BenchmarkPolygonSingleRingValidation/n=10               5276            4300            -18.50%
BenchmarkPolygonSingleRingValidation/n=100              482064          39632           -91.78%
BenchmarkPolygonSingleRingValidation/n=1000             48003617        379272          -99.21%
BenchmarkPolygonSingleRingValidation/n=10000            4801141376      4910661         -99.90%
BenchmarkPolygonMultipleRingsValidation/n=4             27512           30712           +11.63%
BenchmarkPolygonMultipleRingsValidation/n=36            1180281         1203961         +2.01%
BenchmarkPolygonMultipleRingsValidation/n=400           131915204       132171840       +0.19%
BenchmarkPolygonMultipleRingsValidation/n=4096          13704004464     13706626576     +0.02%
```